Issue #128: Extracted allocated mode updaters to dedicated module (`AllocatePageAllocateModeUpdaters.js`).
